+++ CHANGES.md
@@ -11,8 +11,11 @@
 
 ## develop
 
+- [UPDATE] JetPack 4.6 に上げる
+    - @tnoho
 - [UPDATE] `libwebrtc` を `97.4692.4.0` に上げる
     - @voluntas
+
 
 ## 2021.6.0