+++ CHANGES.md
@@ -11,6 +11,14 @@
 
 ## develop
 
+- [UPDATE] Raspberry Pi OS bullseye に対応
+    - @tnoho
+- [UPDATE] JetPack 4.6 に上げる
+    - @tnoho
+- [UPDATE] `libwebrtc` を `97.4692.4.0` に上げる
+    - @voluntas
+
+
 ## 2021.6.0
 
 - [UPDATE] Boost 1.78.0 に上げる