+++ CHANGES.md
@@ -23,6 +23,8 @@
     - @voluntas
 - [FIX] スクリーンキャプチャが Linux で落ちるのを修正
     - @tnoho
+- [ADD] DataChannel を使うことになっていて Offer を行う際には DataChannel を作るように変更
+    - @tnoho
 
 
 ## 2021.6.0