+++ CHANGES.md
@@ -23,6 +23,10 @@
     - @voluntas
 - [FIX] Jetson のハードウェアデコーダーが出力時に出力サイズでフレームを切り抜いていなかったため修正
     - @tnoho
+- [FIX] スクリーンキャプチャが Linux で落ちるのを修正
+    - @tnoho
+- [ADD] DataChannel を使うことになっていて Offer を行う際には DataChannel を作るように変更
+    - @tnoho
 
 
 ## 2021.6.0